Remove mention of validate_drp and rename
This renames all the jobs that used to be called validate_drp
to verify_drp_metrics
