+++ pipelines/release/test_official_release.groovy
@@ -221,16 +221,16 @@ notify.wrap {
       } // retry
     }
 
-    triggerMe['validate_drp'] = {
+    triggerMe['verify_drp_metrics'] = {
       retry(1) {
         // based on lsstsqre/stack image
         build(
-          job: 'sqre/validate_drp',
+          job: 'sqre/verify_drp_metrics',
           parameters: [
             string(name: 'DOCKER_IMAGE', value: stackResults.image),
             booleanParam(
               name: 'NO_PUSH',
-              value: scipipe.release.step.validate_drp.no_push,
+              value: scipipe.release.step.verify_drp_metrics.no_push,
             ),
             booleanParam(name: 'WIPEOUT', value: false),
           ],
